# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: longue-vue CMDB API
  version: 0.1.0
  summary: Kubernetes-centric CMDB aligned with the ANSSI SecNumCloud framework.
  description: |
    REST API for **longue-vue**, a CMDB for Kubernetes environments aligned with the
    ANSSI SecNumCloud (SNC) qualification framework.

    This version (0.1.0) is a walking skeleton covering:
      - Liveness and readiness probes
      - `Cluster` resource CRUD

    Subsequent versions will add Node, Namespace, Workload, Pod, and other
    Kubernetes entity kinds mapped onto the ANSSI cartography layers.

    Errors follow RFC 7807 (`application/problem+json`).

    Interactive API documentation is available at `/docs/` for authenticated
    operators; sign in via `/ui/login` first to use "Try it out".
servers:
  - url: /
    description: Current origin (recommended for browser-based "Try it out" via /docs/)
  - url: http://localhost:8080
    description: Local development
...
security:
  - BearerAuth: []
  - SessionCookie: []
paths:
...
  /v1/workloads:
    get:
      tags: [workloads]
      summary: List workloads
      operationId: listWorkloads
      security:
        - BearerAuth: [read]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/IncludeTerminated'
        - $ref: '#/components/parameters/WorkloadNamespaceIdFilter'
        - $ref: '#/components/parameters/WorkloadKindFilter'
        - $ref: '#/components/parameters/WorkloadImageFilter'
        - $ref: '#/components/parameters/WorkloadApplicationIdFilter'
        - $ref: '#/components/parameters/WorkloadApplicationNameFilter'
        - $ref: '#/components/parameters/WorkloadUnlinkedFilter'
      responses:
        '200':
          description: Paged list of workloads.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkloadList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [workloads]
      summary: Register a workload
      operationId: createWorkload
      security:
        - BearerAuth: [write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkloadCreate'
      responses:
        '201':
          description: Workload created.
          headers:
            Location:
              description: URI of the newly created workload.
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workload'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /v1/applications:
    get:
      tags: [applications]
      summary: List applications
      operationId: listApplications
      security:
        - BearerAuth: [read]
        - SessionCookie: [read]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: name
          in: query
          required: false
          description: Case-insensitive substring filter on name + display_name.
          schema: { type: string }
        - name: application_block_id
          in: query
          required: false
          description: Restrict to applications belonging to this block.
          schema:
            type: string
            format: uuid
        - name: application_block_name
          in: query
          required: false
          description: Restrict to applications belonging to the block with this exact name.
          schema: { type: string }
        - name: criticality
          in: query
          required: false
          description: Exact criticality filter.
          schema: { type: string }
        - name: has_dict
          in: query
          required: false
          description: |
            When true, restrict to applications with at least one DICT
            classification field set. When false, restrict to applications
            with none set.
          schema: { type: boolean }
        - name: dict_min
          in: query
          required: false
          description: |
            Restrict to applications whose `MAX(sec_*)` is at least this
            value (SNC severity floor filter).
          schema:
            type: integer
            minimum: 0
            maximum: 4
      responses:
        '200':
          description: Paged list of applications.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [applications]
      summary: Create an application
      description: |
        Either `application_block_id` (UUID) or `application_block_name`
        (string) may be supplied to link the new application to a block.
        If both are present the UUID wins.
      operationId: createApplication
      security:
        - BearerAuth: [write]
        - SessionCookie: [write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationCreate'
      responses:
        '201':
          description: Application created.
          headers:
            Location:
              description: URI of the newly created application.
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /v1/applications/{id}:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
    get:
      tags: [applications]
      summary: Get an application
      operationId: getApplication
      security:
        - BearerAuth: [read]
        - SessionCookie: [read]
      responses:
        '200':
          description: Application details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [applications]
      summary: Update mutable fields of an application
      description: |
        Merge-patch semantics: fields omitted from the body are unchanged.
        `name` is immutable after creation. Supplying either
        `application_block_id` or `application_block_name` re-points the
        application to a different block.
      operationId: updateApplication
      security:
        - BearerAuth: [write]
        - SessionCookie: [write]
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/ApplicationPatch'
      responses:
        '200':
          description: Updated application.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [applications]
      summary: Delete an application
      description: |
        Deletes the application; any workloads or VMs still referencing it
        have their `application_id` FK reset to NULL.
      operationId: deleteApplication
      security:
        - BearerAuth: [delete]
        - SessionCookie: [delete]
      responses:
        '204':
          description: Application deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/applications/by-name/{name}:
    parameters:
      - name: name
        in: path
        required: true
        description: Normalised application name (see ADR-0029 §3).
        schema:
          type: string
    get:
      tags: [applications]
      summary: Get an application by name
      operationId: getApplicationByName
      security:
        - BearerAuth: [read]
        - SessionCookie: [read]
      responses:
        '200':
          description: Application details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/applications/{id}/members:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
    get:
      tags: [applications]
      summary: List the members of an application
      description: |
        Returns every workload, virtual machine and VM application linked
        to this application, with a per-row `linked_at` / `linked_by`
        provenance triple.
      operationId: listApplicationMembers
      security:
        - BearerAuth: [read]
        - SessionCookie: [read]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: kind
          in: query
          required: false
          description: Restrict to members of this kind.
          schema:
            type: string
            enum: [workload, virtual_machine, vm_application]
      responses:
        '200':
          description: Paged list of application members.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationMemberList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: |
        API token issued out-of-band. Send as `Authorization: Bearer <token>`.

        Scopes granted to a token:
          - `read`   — list and get operations
          - `write`  — create and update (implies `read` at the caller's discretion)
          - `delete` — remove resources
          - `admin`  — all of the above; reserved for future admin-only endpoints

        Tokens carrying the `admin` scope implicitly satisfy any scope required
        by an operation.

        Per ADR-0007, machine tokens are created by an admin in the UI
        (see `POST /v1/admin/tokens`) rather than injected through env vars.

    SessionCookie:
      type: apiKey
      in: cookie
      name: longue_vue_session
      description: |
        Human session cookie set by `POST /v1/auth/login`. HttpOnly,
        SameSite=Strict, 8-hour sliding expiry. Per ADR-0007.

        Every operation accepts either `BearerAuth` or `SessionCookie` —
        scope enforcement reads from the caller's role (for sessions) or
        token scopes (for bearer), so the two paths share the same
        downstream authorisation check.
  parameters:
...
    WorkloadNamespaceIdFilter:
      name: namespace_id
      in: query
      required: false
      description: Return only workloads belonging to this namespace.
      schema:
        type: string
        format: uuid
    WorkloadKindFilter:
      name: kind
      in: query
      required: false
      description: Return only workloads of this kind.
      schema:
        $ref: '#/components/schemas/WorkloadKind'
    WorkloadImageFilter:
      name: image
      in: query
      required: false
      description: |
        Case-insensitive substring match against any container's `image`
        field in the workload's containers JSON (init containers included).
        Mirrors the PodImageFilter so both kinds answer the same
        "which assets run this component version?" question.
      schema:
        type: string
        maxLength: 253
    WorkloadApplicationIdFilter:
      name: application_id
      in: query
      required: false
      description: Return only workloads linked to this Application (ADR-0029).
      schema:
        type: string
        format: uuid
    WorkloadApplicationNameFilter:
      name: application_name
      in: query
      required: false
      description: |
        Return only workloads linked to the Application with this name
        (ADR-0029). Resolved server-side.
      schema:
        type: string
    WorkloadUnlinkedFilter:
      name: unlinked
      in: query
      required: false
      description: When true, return only workloads with application_id IS NULL.
      schema:
        type: boolean
...
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of items to return. Server clamps to [1, 200].
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned from a previous list response.
      schema:
        type: string
        maxLength: 512
    IncludeTerminated:
      name: include_terminated
      in: query
      required: false
      description: |
        When true, include rows whose `terminated_at` is set (soft-deleted by
        the collector reconcile pass). Defaults to false: only live entities
        are returned. ADR-0021 §5 / §6.
      schema:
        type: boolean
        default: false
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Token is valid but lacks the scope required by the operation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    WorkloadCreate:
      description: Payload to register a new workload under a namespace.
      allOf:
        - $ref: '#/components/schemas/WorkloadMutable'
        - type: object
          required: [namespace_id, kind, name]
          properties:
            namespace_id:
              type: string
              format: uuid
              description: |
                Parent namespace id. Immutable after creation; the namespace
                must already exist or the create returns 404.
            kind:
              allOf:
                - $ref: '#/components/schemas/WorkloadKind'
              description: |
                Workload kind discriminator. Immutable after creation.
            name:
              type: string
              description: |
                Kubernetes workload name (DNS-subdomain style). Unique per
                (namespace, kind). Immutable after creation.
              pattern: '^[a-z0-9]([-a-z0-9.]*[a-z0-9])?$'
              minLength: 1
              maxLength: 253
...
    WorkloadList:
      description: Paged list of workloads.
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Workload'
        next_cursor:
          type: string
          nullable: true
          description: |
            Opaque cursor to pass as `?cursor=` to fetch the next page.
            Absent or null when no more pages remain.
...